survery builder data types renamed
